+++ examples/editorial activities/make_enum.py
@@ -1,41 +1,41 @@
 #!/usr/bin/env/python
 
 """
-    make_enum.py -- make enumerations for publications
+    make_enum.py -- make enumerations for editorial activities
 """
+
+from datetime import datetime
+from pump.vivopump import get_parms, create_enum
 
 __author__ = "Michael Conlon"
 __copyright__ = "Copyright 2016 (c) Michael Conlon"
 __license__ = "BSD 3-Clause license"
 __version__ = "0.1.1"
 
-from datetime import datetime
-
-from pump.vivopump import get_parms, create_enum
-
 
 def main():
     """
-    Generate the enums for publications
+    Generate the enums for editorial activities
     """
     print datetime.now(), "Start"
     parms = get_parms()
 
-    #   person via Orcid
+    #   person via label
 
     query = """
     SELECT (MIN (?xshort) AS ?short) ?vivo
     WHERE
     {
-          ?vivo vivo:orcidId ?xshort .
+          ?vivo a foaf:Person .
+          ?vivo rdfs:label ?xshort .
     }
     GROUP BY ?vivo
     ORDER BY ?short
     """
 
-    create_enum("orcid_enum.txt", query, parms)
+    create_enum("person_enum.txt", query, parms)
 
-    #   journals via issn
+    #   journals via label
 
     query = """
     SELECT (MIN (?xlabel) AS ?short) ?vivo
@@ -57,7 +57,6 @@ def main():
     WHERE
     {
           ?vivo a vivo:DateTimeValue .
-#          ?vivo vivo:dateTimePrecision vivo:yearMonthDayPrecision .
           ?vivo vivo:dateTime ?short .
     }
     ORDER BY ?short